mysql_pconnect should reconnect in all error cases
If mysql_ping fails, and the error code was not 2006 (CR_SERVER_GONE_ERROR),
mysql_pconnect would still reuse the connecting, causing subsequent queries to
fail.

If for instance, a PHP script ran out of memory while fetching from a MySQL
result, the reused MySQL connecting would be in an invalid state, causing
queries (and ping) to fail with 2014 (CR_COMMANDS_OUT_OF_SYNC).

Just bypassing the check for error code 2006 would cause a crash, since PHP
would attempt to use the mysql handle directly, so we close the handle and
create a new one.